requirements.txt
Flask==1.1.2
Flask-Bootstrap4==4.0.2
SQLAlchemy==1.3.19
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
python-dotenv==0.14.0